django-backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.2.0
bcrypt==3.2.0
blessed==1.17.10
boto3==1.16.39
botocore==1.19.39
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.0
cement==2.8.2
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1.1
distlib==0.3.1
Django==3.1.4
django-cors-headers==3.5.0
django-dotenv==1.4.2
django-rest-multiple-models==2.1.3
django-storages==1.10.1
django-webpack-loader==0.7.0
djangorestframework==3.12.2
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
fcm-django==0.3.7
filelock==3.0.12
firebase-admin==4.5.0
future==0.16.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
jmespath==0.10.0
jsonschema==3.2.0
msgpack==1.0.2
paramiko==2.7.2
pathspec==0.5.9
pbr==5.5.1
Pillow==8.0.1
pipenv==2020.11.15
pipenv-to-requirements==0.9.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.9
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
s3transfer==0.3.3
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
wcwidth==0.1.9
websocket-client==0.57.0
psycopg2==2.8.4
